RSL-46: refactor code
