+++ src/components/DeleteIconButton/index.js
@@ -1,0 +1,3 @@
+import DeleteIconButton from './DeleteIconButton';
+
+export default DeleteIconButton;
