+++ src/components/DictionaryTable/TablePaginationActions/index.js
@@ -1,0 +1,3 @@
+import TablePaginationActions from './TablePaginationActions';
+
+export default TablePaginationActions;
