+++ src/constants/variables-learning.js
@@ -25,3 +25,19 @@ export const VOTE_BUTTON = [
 ];
 
 export const LEARN_RATINGS = { bad: 20, normal: 50, good: 70, excellent: 90 };
+
+export const DEFAULT_WORD = {
+  optional: {
+    word: 'acre',
+    wordTranslate: 'акр',
+    transcription: '[éikər]',
+    textMeaning: 'An <i>acre</i> is a unit for measuring area.',
+    textMeaningTranslate: 'Акр - это единица измерения площади',
+    textExample: 'They lived on a 150-<b>acre</b> farm.',
+    textExampleTranslate: 'Они жили на 150-акровой ферме',
+    image: 'files/01_1201.jpg',
+    audio: 'files/01_1201.mp3',
+    audioMeaning: 'files/01_1201_meaning.mp3',
+    audioExample: 'files/01_1201_example.mp3',
+  },
+};
